treewide: make clippy clean so CI can run it with -D warnings
Prerequisite for the code CI job; kept separate so style changes do not
mix with workflow configuration. Every warning is fixed on its merits
rather than silenced, and behaviour is unchanged - guarded by the SHA-256
RFC vectors, 56 tests, the fuzzer and two QEMU runs.

Host crates: rewrite the SHA-256 message-schedule load as
zip(chunks_exact(4)) without try_into().unwrap(), since a trusted-base
primitive should not gain even a theoretical panic path; regroup the
version literal as 0x00_02_01 (major_minor_patch, same value); two
match-to-bool into matches!; % MEM_DESC_SIZE into !is_multiple_of; COM1 +
0 into named ports with a note that DLL aliases THR while DLAB is set.

Target crates: RESULT_PORT cast to its own type, (len + 0xfff) / 0x1000
into div_ceil, desc_size % 8 into is_multiple_of, a redundant import.

Two clippy::not_unsafe_ptr_arg_deref findings are deny-by-default errors
on efi_main and boot_entry. Neither can be an unsafe fn: firmware and the
loader enter them through an ABI, not through a Rust call. They get a
targeted #[allow] carrying an actual safety statement - who supplies the
pointer, under which specification clause, and what is validated before
the first dereference. Annotating all 52 unsafe sites remains a Priority
4 item.

Also learned for the CI job: --all-targets must not be used on
tos-uefi-loader or tos-nucleus, because building their test harness pulls
in std and fails with a duplicate panic_impl lang item.

Verified: cargo clippy --all-targets -- -D warnings clean, and clean for
both target triples; cargo test 56 passed / 0 failed; both release builds
succeed; fuzz 50000 PASS; QEMU success PASS (exit 33) and QEMU negative
with invalid-kind-none.bin PASS (exit 67).

+++ source/nucleus/src/main.rs
@@ -18,7 +18,6 @@ use tos_boot_protocol::{
 };
 use tos_capsule::parse;
 use tos_hash::{Sha256, sha256};
-use tos_serial;
 
 #[panic_handler]
 fn panic(_info: &PanicInfo) -> ! {
@@ -32,7 +31,7 @@ fn result_port(code: u8) -> ! {
         asm!(
             "out dx, al",
             in("al") code,
-            in("dx") (RESULT_PORT as u16),
+            in("dx") RESULT_PORT,
             options(nomem, nostack, preserves_flags)
         );
     }
@@ -81,6 +80,13 @@ fn first_logical_line(content: &[u8]) -> Option<&[u8]> {
     None
 }
 
+// SAFETY (entry-point contract): `bi_raw` is the physical address of a BootInfo
+// record placed in `rdi` by the loader per BOOT_ABI_V1 §3, in an identity-mapped
+// region the loader marked reserved. It is validated as raw bytes
+// (`BootInfo::validate_bytes`) before it is ever read as a struct. The function
+// cannot be an `unsafe fn`: the loader transfers control to it with a machine
+// `call`, not a Rust call, so `clippy::not_unsafe_ptr_arg_deref` does not apply.
+#[allow(clippy::not_unsafe_ptr_arg_deref)]
 #[no_mangle]
 #[link_section = ".text.boot_entry"]
 pub extern "C" fn boot_entry(bi_raw: *const BootInfo) -> ! {
